+++ src/main/java/com/edu/service/impl/DownloadServiceImpl.java
@@ -4,6 +4,8 @@ import com.edu.bean.DownloadBean;
 import com.edu.dao.Dao;
 import com.edu.service.DownloadService;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -12,6 +14,8 @@ import java.util.List;
  */
 public class DownloadServiceImpl implements DownloadService {
     private Dao dao = new Dao();
+    private SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+
 
     @Override
     public List<DownloadBean> selectAll() {
@@ -30,8 +34,33 @@ public class DownloadServiceImpl implements DownloadService {
 
     @Override
     public Boolean deleteBySongId(Integer songId) {
-        String sql = "delete from download where song_id = '"+songId+"'";
-        Boolean flag = dao.addObj(sql);
+        String selectSql = "select * from download where song_id = '"+songId+"'";
+        List<DownloadBean> list = (List<DownloadBean>) dao.query(selectSql, DownloadBean.class);
+
+        Boolean flag = false;
+        if (list.size()>0){
+            String sql = "delete from download where song_id = '"+songId+"'";
+            flag = dao.addObj(sql);
+        }else {
+            flag = true;
+        }
+
+        return flag;
+    }
+
+    @Override
+    public Boolean download(Integer userId, Integer songId) {
+        String songSql = "update song set song_download = song_download + 1 where song_id = '"+songId+"'";
+        String clickSql = "insert into download (user_id,song_id,download_date) " +
+                "values ('"+userId+"','"+songId+"','"+simpleDateFormat.format(new Date())+"')";
+
+        Boolean songFlag = dao.addObj(songSql);
+        Boolean clickFlag = dao.addObj(clickSql);
+
+        Boolean flag = false;
+        if (songFlag&&clickFlag){
+            flag = true;
+        }
         return flag;
     }
 }
